+++ test/kube2e/gateway/gateway_suite_test.go
@@ -2,12 +2,11 @@ package gateway_test
 
 import (
 	"context"
+	"github.com/solo-io/gloo/pkg/utils/kubeutils/kubectl"
 	"os"
 	"path/filepath"
 	"testing"
 	"time"
-
-	"github.com/solo-io/gloo/pkg/cliutil/kubectl"
 
 	"github.com/solo-io/gloo/test/testutils/kubeutils"
